Avoid data class in jpa model.
