--- src/main/kotlin/no/nav/klage/dokument/api/input/BrevMottakerInput.kt
@@ -1,9 +1,0 @@
-package no.nav.klage.dokument.api.input
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-data class BrevMottakerInput(
-    val partId: PartIdInput,
-    val navn: String?,
-)
