--- src/main/kotlin/no/nav/klage/dokument/api/input/PartIdInput.kt
@@ -1,3 +1,0 @@
-package no.nav.klage.dokument.api.input
-
-data class PartIdInput(val type: String?, val partIdTypeId: String?, val value: String)
