--- src/main/kotlin/no/nav/klage/dokument/api/input/TilleggsopplysningInput.kt
@@ -1,3 +1,0 @@
-package no.nav.klage.dokument.api.input
-
-data class TilleggsopplysningInput(val key: String, val value: String)
