+++ src/main/kotlin/no/nav/klage/dokument/api/mapper/DokumentEnhetInputMapper.kt
@@ -1,9 +1,7 @@
 package no.nav.klage.dokument.api.mapper
 
-import no.nav.klage.dokument.api.input.BrevMottakerInput
-import no.nav.klage.dokument.api.input.DokumentInput
-import no.nav.klage.dokument.api.input.JournalfoeringDataInput
-import no.nav.klage.dokument.api.input.PartIdInput
+
+import no.nav.klage.dokument.api.input.DokumentEnhetWithDokumentreferanserInput
 import no.nav.klage.dokument.domain.dokument.*
 import no.nav.klage.dokument.exceptions.DokumentEnhetNotValidException
 import no.nav.klage.dokument.util.getLogger
@@ -22,10 +20,10 @@ class DokumentEnhetInputMapper {
         private val secureLogger = getSecureLogger()
     }
 
-    fun mapBrevMottakereInput(brevMottakereInput: List<BrevMottakerInput>): Set<BrevMottaker> =
+    fun mapBrevMottakereInput(brevMottakereInput: List<DokumentEnhetWithDokumentreferanserInput.BrevMottakerInput>): Set<BrevMottaker> =
         brevMottakereInput.map { mapBrevMottakerInput(it) }.toSet()
 
-    fun mapBrevMottakerInput(brevMottakerInput: BrevMottakerInput): BrevMottaker =
+    fun mapBrevMottakerInput(brevMottakerInput: DokumentEnhetWithDokumentreferanserInput.BrevMottakerInput): BrevMottaker =
         try {
             BrevMottaker(
                 partId = mapPartIdInput(brevMottakerInput.partId),
@@ -36,7 +34,7 @@ class DokumentEnhetInputMapper {
             throw DokumentEnhetNotValidException("Ulovlig input: ${iae.message}")
         }
 
-    fun mapJournalfoeringDataInput(input: JournalfoeringDataInput): JournalfoeringData =
+    fun mapJournalfoeringDataInput(input: DokumentEnhetWithDokumentreferanserInput.JournalfoeringDataInput): JournalfoeringData =
         try {
             JournalfoeringData(
                 sakenGjelder = mapPartIdInput(input.sakenGjelder),
@@ -56,7 +54,7 @@ class DokumentEnhetInputMapper {
             throw DokumentEnhetNotValidException("Ulovlig input: ${iae.message}")
         }
 
-    fun mapDokumentInputToHoveddokument(dokument: DokumentInput.Dokument): OpplastetHoveddokument =
+    fun mapDokumentInputToHoveddokument(dokument: DokumentEnhetWithDokumentreferanserInput.DokumentInput.Dokument): OpplastetHoveddokument =
         OpplastetHoveddokument(
             mellomlagerId = dokument.mellomlagerId,
             opplastet = dokument.opplastet,
@@ -64,7 +62,7 @@ class DokumentEnhetInputMapper {
             name = dokument.name
         )
 
-    fun mapDokumentInputToVedlegg(dokument: DokumentInput.Dokument): OpplastetVedlegg =
+    fun mapDokumentInputToVedlegg(dokument: DokumentEnhetWithDokumentreferanserInput.DokumentInput.Dokument): OpplastetVedlegg =
         OpplastetVedlegg(
             mellomlagerId = dokument.mellomlagerId,
             opplastet = dokument.opplastet,
@@ -72,7 +70,7 @@ class DokumentEnhetInputMapper {
             name = dokument.name
         )
 
-    private fun mapPartIdInput(partIdInput: PartIdInput): PartId {
+    private fun mapPartIdInput(partIdInput: DokumentEnhetWithDokumentreferanserInput.PartIdInput): PartId {
         return try {
             PartId(
                 type = if (partIdInput.partIdTypeId != null) PartIdType.of(partIdInput.partIdTypeId)
