+++ src/main/kotlin/no/nav/klage/dokument/domain/dokument/Tilleggsopplysning.kt
@@ -4,7 +4,7 @@ import jakarta.persistence.Column
 import jakarta.persistence.Embeddable
 
 @Embeddable
-data class Tilleggsopplysning(
+class Tilleggsopplysning(
     @Column(name = "key")
     val key: String,
     @Column(name = "value")
